Moved CryptoError into a separate module.
