+++ components/crypto/src/lib.rs
@@ -13,38 +13,11 @@ extern crate common;
 extern crate serde_derive;
 
 pub mod dh;
+pub mod error;
 pub mod hash;
 pub mod hash_lock;
 pub mod identity;
-// pub mod invoice_id;
-pub mod nonce_window;
-// pub mod payment_id;
+// pub mod nonce_window;
 pub mod rand;
 pub mod sym_encrypt;
 pub mod test_utils;
-// pub mod uid;
-
-use derive_more::*;
-
-#[derive(Clone, Copy, Debug, PartialEq, Display)]
-#[display(fmt = "crypto error")]
-pub struct CryptoError;
-
-impl From<::ring::error::Unspecified> for CryptoError {
-    fn from(_: ::ring::error::Unspecified) -> CryptoError {
-        CryptoError
-    }
-}
-
-/// Increase the bytes represented number by 1.
-///
-/// Reference: `libsodium/sodium/utils.c#L241`
-#[inline]
-pub fn increase_nonce(nonce: &mut [u8]) {
-    let mut c: u16 = 1;
-    for i in nonce {
-        c += u16::from(*i);
-        *i = c as u8;
-        c >>= 8;
-    }
-}
